Create recomemded heating time
A microwave oven manufacturer recommends that when heating two items, add 50% to the heating time, and when heating three items double the heating time. Heating more than three items at once is not recommended. Write a program that asks the user for the number of items and the single-item heating time. The program then writes out the recommended heating time.
Sample Input and Output 1:
Enter the number of items
2
Enter the single item heating time
5.0
The recommended heating time is 7.50